requirements.txt
discord.py==2.4.0
google-generativeai==0.8.3
gTTS==2.5.3
python-dotenv==1.0.1
PyNaCl==1.5.0

